rawdata-server: Adding some constant conversion library for Python
Some constants cannot be defined with the Google protocol buffer library
as they are defined in the Riot API. For example, the maps
correspondance (map id <-> map name) does not have a consecutive order,
which is not supported for the protocol buffer library (or maybe I did
not read enough the documentation).

This short library is here to make the conversion from the map ID to a
protocol buffer enumerator.