rawdata-server: Converter from JSON to protocol buffer
Adding a converter from JSON to protocol buffer. This converter has
tests based on the outputs of the official Riot API [1]. We do not save
all contents send by the Riot API, but mostly the "interesting"
content. If new fields are implemented in the protocol buffer, this
class also need to be modified.

  [1] https://developer.riotgames.com/api/methods#!/1224/4756